Remove bogus cast.
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@1110 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/SOARecord.java
@@ -92,7 +92,7 @@ throws IOException
 	SOARecord rec = new SOARecord(name, dclass, ttl);
 	rec.host = st.getName(origin);
 	rec.admin = st.getName(origin);
-	rec.serial = (int) st.getUInt32();
+	rec.serial = st.getUInt32();
 	rec.refresh = st.getTTL();
 	rec.retry = st.getTTL();
 	rec.expire = st.getTTL();
